work in progress tableView data source
FOLLOWING will be deleted due to data-view sync mechanism change, keep
a record in this commit for future reference
#pragma mark - Sync with local db
/*
 1.     For any row in tableDataSource, search for the counterpart in
refreshed rawDataSource.
 1.1    Search for same serverId
 1.1.1    Match by versionNo to ensure the latest version is shown
 1.1.2    Match by lastModifiedAtLocal to ensure local made changes are
shown
 1.1.3    For those with no match in tableDataSource, delete them.
 1.2    Search for same localId in case that serverId is not available.
versionNo is not able to be used here since it's from server and no
serverId means it has not successfully communicated with server.
 1.2.1    Match by lastModifiedAtLocal to ensure local made changes are
shown
 2. For those in rawDataSource without a match in tableDataSource, add
them to tableDataSource anyway.
 To easily achieve above steps, we create a new array based on
rawDataSource. Add a flag to each of it's object to mark if it is
matched during the search.
 */
- (void)syncWithLocalDb:(NSString *)userServerId